// applications/webapp/src/app/api/generated/model/readiness-result.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReadinessResult { 
    ready?: boolean;
    checks?: { [key: string]: string; };
}